feat : RaceService > 전진 수행 메서드, 우승자 추출 메서드 구현
